Fixed bug when select storage dialog not shows on first start of application
