+++ OsmAnd/src/net/osmand/plus/download/DownloadIndexFragment.java
@@ -105,70 +105,6 @@ public class DownloadIndexFragment extends OsmandExpandableListFragment {
 		settings = getMyApplication().getSettings();
 
 
-		if(getMyApplication().getResourceManager().getIndexFileNames().isEmpty()) {
-			boolean showedDialog = false;
-			if(Build.VERSION.SDK_INT < OsmandSettings.VERSION_DEFAULTLOCATION_CHANGED) {
-				SuggestExternalDirectoryDialog.showDialog(getActivity(), null, null);
-			}
-			if(!showedDialog) {
-				getDownloadActivity().showDialogOfFreeDownloadsIfNeeded();
-			}
-		} else {
-			getDownloadActivity().showDialogOfFreeDownloadsIfNeeded();
-		}
-		
-
-		if (Build.VERSION.SDK_INT >= OsmandSettings.VERSION_DEFAULTLOCATION_CHANGED) {
-			final String currentStorage = settings.getExternalStorageDirectory().getAbsolutePath();
-			String primaryStorage = settings.getDefaultExternalStorageLocation();
-			if (!currentStorage.startsWith(primaryStorage)) {
-				// secondary storage
-				boolean currentDirectoryNotWritable = true;
-				for (String writeableDirectory : settings.getWritableSecondaryStorageDirectorys()) {
-					if (currentStorage.startsWith(writeableDirectory)) {
-						currentDirectoryNotWritable = false;
-						break;
-					}
-				}
-				if (currentDirectoryNotWritable) {
-					currentDirectoryNotWritable = !OsmandSettings.isWritable(settings.getExternalStorageDirectory());
-				}
-				if (currentDirectoryNotWritable) {
-					final String newLoc = settings.getMatchingExternalFilesDir(currentStorage);
-					if (newLoc != null && newLoc.length() != 0) {
-						AccessibleAlertBuilder ab = new AccessibleAlertBuilder(getDownloadActivity());
-						ab.setMessage(getString(R.string.android_19_location_disabled,
-								settings.getExternalStorageDirectory()));
-						ab.setPositiveButton(R.string.default_buttons_yes, new DialogInterface.OnClickListener() {
-							@Override
-							public void onClick(DialogInterface dialog, int which) {
-								copyFilesForAndroid19(newLoc);
-							}
-						});
-						ab.setNegativeButton(R.string.default_buttons_cancel, null);
-						ab.show();
-					}
-				}
-			}
-		}
-	}
-	
-	private void copyFilesForAndroid19(final String newLoc) {
-		MoveFilesToDifferentDirectory task = 
-				new MoveFilesToDifferentDirectory(getDownloadActivity(),
-						new File(settings.getExternalStorageDirectory(), IndexConstants.APP_DIR), 
-						new File(newLoc, IndexConstants.APP_DIR)) {
-			protected Boolean doInBackground(Void[] params) {
-				Boolean result = super.doInBackground(params);
-				if(result) {
-					settings.setExternalStorageDirectory(newLoc);
-					getMyApplication().getResourceManager().resetStoreDirectory();
-					getMyApplication().getResourceManager().reloadIndexes(progress)	;
-				}
-				return result;
-			};
-		};
-		task.execute();
 	}
 	
 	@Override
